feat(module-replacement): rename
