+++ package.json
@@ -52,7 +52,7 @@
     "packages/falcon-dev-tools/babel-preset-falcon-client",
     "packages/falcon-dev-tools/eslint-config-falcon",
     "packages/falcon-dev-tools/falcon-i18n-webpack-plugin",
-    "packages/falcon-dev-tools/falcon-normal-module-override-webpack-plugin",
+    "packages/falcon-dev-tools/normal-module-override-webpack-plugin",
     "packages/eject-ts",
     "examples/shop-with-blog/client",
     "examples/shop-with-blog/server"
